+++ src/main/java/ru/mail/polis/client/AsyncHttpClient.java
@@ -7,7 +7,9 @@ import java.util.concurrent.CompletableFuture;
 
 public interface AsyncHttpClient {
     CompletableFuture<Void> upsert(@NotNull final byte[] value, @NotNull final String id, @NotNull final String url);
+
     CompletableFuture<Void> delete(@NotNull final String id, @NotNull final String url);
+    
     CompletableFuture<Value> get(@NotNull final String id, @NotNull final String url);
 
     static AsyncHttpClient create() {
