+++ src/main/java/ru/mail/polis/promise/CompletablePromise.java
@@ -1,5 +1,7 @@
 package ru.mail.polis.promise;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
@@ -7,7 +9,7 @@ import java.util.concurrent.Future;
 public final class CompletablePromise<V> extends CompletableFuture<V> {
     private final Future<V> future;
 
-    public CompletablePromise(Future<V> future) {
+    public CompletablePromise(@NotNull final Future<V> future) {
         this.future = future;
         CompletablePromiseContext.schedule(this::tryToComplete);
     }
@@ -18,6 +20,7 @@ public final class CompletablePromise<V> extends CompletableFuture<V> {
                 complete(future.get());
             } catch (InterruptedException e) {
                 completeExceptionally(e);
+                Thread.currentThread().interrupt();
             } catch (ExecutionException e) {
                 completeExceptionally(e.getCause());
             }
