+++ src/main/java/ru/mail/polis/promise/CompletablePromiseContext.java
@@ -1,5 +1,7 @@
 package ru.mail.polis.promise;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
@@ -7,11 +9,11 @@ import java.util.concurrent.TimeUnit;
 @SuppressWarnings("FutureReturnValueIgnored")
 final class CompletablePromiseContext {
 
-    private CompletablePromiseContext(){}
-
     private static final ScheduledExecutorService SERVICE = Executors.newSingleThreadScheduledExecutor();
 
-    static void schedule(Runnable r) {
+    private CompletablePromiseContext(){}
+
+    static void schedule(@NotNull final Runnable r) {
         SERVICE.schedule(r, 20, TimeUnit.MILLISECONDS);
     }
 }
